tools: Move sniff_compiler_version to compiler.py
